+++ test/testproject/proxy/odatajs.d.ts
@@ -1,0 +1,14 @@
+declare module "odatajs" {
+    class oData {
+        static request(request: Request, success?: (data: any, response: any) => void, error?: (error: any) => void, handler?, httpClient?, metadata?);
+    }
+
+    interface Request {
+        requestUri: string,
+        method: string,
+        headers: Header | Header[],
+        data?: any
+    }
+
+    interface Header { [name: string]: string }
+}