+++ CHANGELOG.md
@@ -1,24 +1,68 @@
 # Release Notes
 
+All notable changes to the "odatatools" will be documented in this file.
+
+Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+
+## 1.0.0 2017-06-05 - Typescript 2.1+ overhaul
+
+### Added
+
+- Setting: Extension Version selection to be able geting old behavior (0.x)
+- Setting: Insider mode (Not used for anything so far)
+- Async and await ES3/5/6+ Proxy
+- Additional headers for interface generation (only version 1.0+)
+- Additional headers for proxy generation (only version 1.0+)
+- Hook generation in proxy file
+
+### Changed
+
+- Changelog file is now structured as suggested by keep a changelog
+- Hook in interface file is now json format (only version 1.0+)
+- Hooks are now pasted at the start with generation notes (only version 1.0+)
+
+### Deprecated
+
+- Version 4.0 and below custom classes (Thenable caller, etc.)
+- Support for version 4.0 and below
+
+### Removed
+
+- Delta classes generation when generating interfaces (only Version 1.0+)
+
 ## 0.4.0 OData Actions and Functions
 
-- **NEW** OData V4 client generator can now handle bound and unbound actions and Functions
-- **NEW** OData V4 client allows custom headers in constructor (experimental - not tested)
-- **IMPROVEMENT** OData V4 Client generator takes now namespace and container name for generated proxy
+### Added
+
+- OData V4 client generator can now handle bound and unbound actions and Functions
+- OData V4 client allows custom headers in constructor (experimental - not tested)
+
+### Changed
+
+- OData V4 Client generator takes now namespace and container name for generated proxy
 
 ## 0.3.0 OData Client
 
-- **NEW** OData V4 client generator (experimental)
-- **FIX** Enums are now string Enums
-- **FIX** Errors when creating interfaces are now caught and printed out
+### Added
+
+- OData V4 client generator (experimental)
+
+### Fixed
+
+- Enums are now string Enums
+- Errors when creating interfaces are now caught and printed out
 
 ### 0.3.1 Hotfix
 
-- **FIX** OData Client switched name and address in constructor
+### Fixed
+
+- OData Client switched name and address in constructor
 
 ## 0.2.0 Ambient and modular declarations
 
-- **FIX** d.ts files will now cause `declare` statements before `namespace`
+### Fixed
+
+- d.ts files will now cause `declare` statements before `namespace`
 
 ## 0.1.0 Interface Generator